requirements.txt
streamlit
langchain
langchain-community
langchain-core
chromadb
pypdf
sentence-transformers
python-dotenv
groq